Refactor CLI Commands (#73)
