+++ package.json
@@ -20,7 +20,7 @@
   },
   "main": "./build/src",
   "bin": {
-    "bluehawk": "./build/src/cli/main.js"
+    "bluehawk": "./build/src/main.js"
   },
   "scripts": {
     "clean": "rm -rf build",
