+++ src/bluehawk/actions/ActionArgs.ts
@@ -1,0 +1,3 @@
+export interface ActionArgs {
+  waitForListeners?: boolean;
+}
