+++ src/bluehawk/actions/check.ts
@@ -1,17 +1,16 @@
-import { CommandModule, Arguments } from "yargs";
 import { getBluehawk } from "../../bluehawk";
-import { withIgnoreOption, withJsonOption } from "../options";
-import { MainArgs } from "../cli";
 import { BluehawkError } from "../../bluehawk/BluehawkError";
-import { printJsonResult } from "../printJsonResult";
 import { logErrorsToConsole } from "../../bluehawk/OnErrorFunction";
+import { ActionArgs } from "./ActionArgs";
+import { printJsonResult } from "./printJsonResult";
 
-interface CheckArgs extends MainArgs {
+export interface CheckArgs extends ActionArgs {
   paths: string[];
   ignore?: string | string[];
+  json?: boolean;
 }
 
-export const check = async (args: Arguments<CheckArgs>): Promise<void> => {
+export const check = async (args: CheckArgs): Promise<void> => {
   const { ignore, json, paths, waitForListeners } = args;
   const bluehawk = await getBluehawk();
   const fileToErrorMap = new Map<string, BluehawkError[]>();
@@ -44,19 +43,3 @@ export const check = async (args: Arguments<CheckArgs>): Promise<void> => {
 
   process.exit(fileToErrorMap.size);
 };
-
-const commandModule: CommandModule<MainArgs & { paths: string[] }, CheckArgs> =
-  {
-    command: "check <paths..>",
-    builder(yargs) {
-      return withJsonOption(withIgnoreOption(yargs));
-    },
-    async handler(args) {
-      return await check(args);
-    },
-    aliases: [],
-    describe:
-      "validate bluehawk markup without outputting any files. Exits with non-zero code if errors are found.",
-  };
-
-export default commandModule;
