+++ src/bluehawk/actions/copy.ts
@@ -1,17 +1,11 @@
 import { Stats } from "fs";
 import * as path from "path";
-import { CommandModule, Arguments, Argv } from "yargs";
 import { getBluehawk } from "../../bluehawk";
-import { MainArgs } from "../cli";
-import {
-  withDestinationOption,
-  withStateOption,
-  withIgnoreOption,
-} from "../options";
+import { ActionArgs } from "./ActionArgs";
 import { System } from "../../bluehawk/io/System";
 import { logErrorsToConsole } from "../../bluehawk/OnErrorFunction";
 
-export interface CopyArgs extends MainArgs {
+export interface CopyArgs extends ActionArgs {
   rootPath: string;
   destination: string;
   state?: string;
@@ -139,35 +133,11 @@ export const copy = async (args: CopyArgs): Promise<string[]> => {
   return errors;
 };
 
-const commandModule: CommandModule<MainArgs & { rootPath: string }, CopyArgs> =
-  {
-    command: "copy <rootPath>",
-    builder: (yargs): Argv<CopyArgs> => {
-      return withIgnoreOption(withStateOption(withDestinationOption(yargs)));
-    },
-    handler: async (args: Arguments<CopyArgs>) => {
-      const errors = await copy(args);
-      if (errors.length !== 0) {
-        console.error(
-          `Exiting with ${errors.length} error${
-            errors.length === 1 ? "" : "s"
-          }.`
-        );
-        process.exit(1);
-      }
-    },
-    aliases: [],
-    describe:
-      "clone source project to destination with Bluehawk commands processed",
-  };
-
-export default commandModule;
-
 /**
   Copy permissions (using stat/chmod) to the given path from the file at the
   given path.
  */
-export const copyPermissions = async ({
+const copyPermissions = async ({
   to,
   from,
 }: {
