+++ src/bluehawk/actions/listCommands.ts
@@ -1,12 +1,12 @@
-import { Arguments, CommandModule } from "yargs";
-import { getBluehawk } from "../../../bluehawk";
-import { MainArgs } from "../../cli";
-import { withJsonOption } from "../../options";
-import { printJsonResult } from "../../printJsonResult";
+import { getBluehawk } from "../../bluehawk";
+import { ActionArgs } from "./ActionArgs";
+import { printJsonResult } from "./printJsonResult";
 
-const handler = async (
-  args: Arguments<MainArgs & { json?: boolean }>
-): Promise<void> => {
+export interface ListCommandArgs extends ActionArgs {
+  json?: boolean;
+}
+
+export const listCommands = async (args: ListCommandArgs): Promise<void> => {
   const { json } = args;
   const bluehawk = await getBluehawk();
 
@@ -62,15 +62,3 @@ const handler = async (
 
   console.log(`available markup commands:\n\n${commandsListText}`);
 };
-
-const commandModule: CommandModule<MainArgs, MainArgs & { json?: boolean }> = {
-  command: "commands",
-  builder(args) {
-    return withJsonOption(args);
-  },
-  handler,
-  aliases: [],
-  describe: "list available commands",
-};
-
-export default commandModule;
