+++ src/bluehawk/actions/snip.test.ts
