+++ src/cli/commandModules/index.ts
@@ -1,5 +1,4 @@
 export * from "./check";
 export * from "./copy";
-export * from "./list/";
 export * from "./list";
 export * from "./snip";
