+++ src/cli/index.ts
@@ -1,5 +1,1 @@
-export * from "./Plugin";
-export * from "./cli";
-export * from "./commands";
-export * from "./options";
-export * from "./printJsonResult";
+export { run } from "./cli";
