+++ src/index.ts
@@ -1,2 +1,1 @@
-export * from "./cli";
 export * from "./bluehawk";
